#389 refactoring, renaming source reference to nested reference for reuse in target
